move configuration directories out of root.
