+++ .gitignore
@@ -1,4 +1,4 @@
 /node_modules/
 /public/build/
-
+/docs/notes.md
 .DS_Store
