+++ package.json
@@ -3,12 +3,12 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "rollup": "rollup -c -w",
+    "rollup": "rollup -c ./cfg/rollup/rollup.config.js -w",
     "tauri": "tauri",
-    "build": "rollup -c",
-    "dev": "concurrently \"rollup -c -w\" \"npm run tauri dev\"",
+    "build": "rollup -c ./cfg/rollup/rollup.config.js",
+    "dev": "concurrently \"rollup -c  ./cfg/rollup/rollup.config.js -w\" \"npm run tauri dev\"",
     "start": "sirv public --no-clear",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
+    "check": "svelte-check --tsconfig .cfg/typescript/tsconfig.json"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
