+++ src-wasm/Cargo.toml.d.ts
@@ -1,3 +1,11 @@
+/** I use this interface so that when I import the Rust library code through typescript,
+ *  I get the interface through which I can call webassembly code */
+
+/** feedback type for letters.
+ * 0 represents a letter which isn't found.
+ * 1 represents a letter which is an exact match.
+ * 2 represents a letter which is in the incorrect spot.
+ */
 type Feedback = 0 | 1 | 2
 
 declare interface Dictionary {
@@ -22,4 +30,5 @@ declare interface Dictionary {
     feedback(guess_word, secret_word): Feedback[]
 }
 
+// when you import the web-assembly library, you typically import a function which returns a promise to the API.
 export default function wasm(params: object): Promise<Dictionary>
