+++ src/apps/Game.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-	import wasm, { Dictionary } from "../src-wasm/Cargo.toml";
-  import {arrayEq} from "./utils";
+	import wasm, { Dictionary, Feedback } from "../../src-wasm/Cargo.toml";
+  import {arrayEq} from "../etc/utils";
+  
   // wordle table size
   const ROWS = 6;
   const COLS = 5;
@@ -8,12 +9,20 @@
   // regular expression for matching a letter
   const REG = /^[A-Z]$/;
 
-  // stores the data in the table
-  let content = Array(ROWS).fill(0).map(()=>Array(COLS).fill(""));
-  let feedback_arr = [];
-  let letter_nums = 0 /* current col idx */
-  let guesses = 0 /* current row idx */
+  // need to make a ROW by COL matrix for the data table
+  /** array corresponding to the rows of the wordle table. Each row is an array of 5 letters. */
+  let content: string[][] = Array(ROWS).fill(0).map(()=>Array(COLS).fill(''));
+
+  /** contains the feedback for previous guesses. */
+  let feedback_arr: Feedback[][] = [];
+  
+  /** number of letters written down in the current row */
+  let letter_nums = 0
+  /** number of previous guesses, which also acts as an index into the current row. */
+  let guesses = 0
+  /** generated by web-assembly code. */
   let secret_word: string;
+  
   // import wasm code
   let wasm_loaded = false;
   let wasm_exports: Dictionary | null = null;
@@ -100,9 +109,7 @@
     }
   }
 
-
   function handle_key(ev) { 
-
     // match the keycode
     switch (ev.keyCode) {
       case 8: {
