+++ src/apps/global.d.ts
