--- src/home.svelte
@@ -1,3 +1,0 @@
-<script>
-  
-</script>
